// src/domain/calidad/models/CalidadPorcentajeGeneralDanosUpdate.ts
export interface CalidadPorcentajeGeneralDanosUpdate {
  IdPorcentajeGeneralDanos: number;
  Trips: string;
  Rona: string;
  Golpe: string;
  Rozadura: string;
  Varicela: string;
  Clavo: string;
  Alga: string;
  Escama: string;
  Granizo: string;
  Gusano: string;
  Quemadura: string;
  Sunblotch: string;
}